interactive_rect_grob drawing

draw an interactive_rect_grob

# S3 method for interactive_rect_grob
drawDetails(x, recording)

Arguments

x
A grid grob.

recording
A logical value indicating whether a grob is being added to the display list or redrawn from the display list.