SVG Graphics Driver

This function produces SVG files (compliant to the current w3 svg XML standard) where elements can be made interactive.

dsvg(file = "Rplots.svg", width = 6, height = 6, bg = "white",
  pointsize = 12, standalone = TRUE, canvas_id = "svg_1",
  fonts = list())

Arguments

file
the file where output will appear.

height, width
Height and width in inches.

bg
Default background color for the plot (defaults to "white").

pointsize
default point size.

standalone
Produce a stand alone svg file? If FALSE, omits xml header and default namespace.

canvas_id
svg id within HTML page.

fonts
Named list of font names to be aliased with fonts installed on your system. If unspecified, the R default families sans, serif, mono and symbol are aliased to the family returned by match_family().

See also

Examples

dsvg()
plot(rnorm(10), main="Simple Example", xlab = "", ylab = "")
dev.off()

#> pdf 
#>   2