interactive points

The geometry is based on geom_point. See the documentation for those functions for more details.

geom_point_interactive(mapping = NULL, data = NULL,
  stat = "identity", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)

Arguments

mapping
The aesthetic mapping, see geom_point.

data
A data frame, see geom_point.

stat
The statistical transformation to use on the data for this layer, as a string, see geom_point.

position
Postion adjustment, see geom_point.

na.rm
See geom_point.

show.legend
See geom_point.

inherit.aes
See geom_point.

...
other arguments passed on to layer. See geom_point.

Note

The following shapes id 3, 4 and 7 to 14 are composite symbols and should not be used.

See also

Examples

# add interactive points to a ggplot -------
library(ggplot2)

dataset <- structure(list(qsec = c(16.46, 17.02, 18.61, 19.44, 17.02, 20.22
), disp = c(160, 160, 108, 258, 360, 225), carname = c("Mazda RX4",
"Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive", "Hornet Sportabout",
"Valiant"), wt = c(2.62, 2.875, 2.32, 3.215, 3.44, 3.46)), row.names = c("Mazda RX4",
"Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive", "Hornet Sportabout",
"Valiant"), class = "data.frame")
dataset

#>                    qsec disp           carname    wt
#> Mazda RX4         16.46  160         Mazda RX4 2.620
#> Mazda RX4 Wag     17.02  160     Mazda RX4 Wag 2.875
#> Datsun 710        18.61  108        Datsun 710 2.320
#> Hornet 4 Drive    19.44  258    Hornet 4 Drive 3.215
#> Hornet Sportabout 17.02  360 Hornet Sportabout 3.440
#> Valiant           20.22  225           Valiant 3.460

# plots
gg_point = ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
        geom_point_interactive() + theme_minimal()

girafe(ggobj = gg_point)