interactive polygons

The geometry is based on geom_polygon. See the documentation for those functions for more details.

geom_polygon_interactive(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)

Arguments

mapping
The aesthetic mapping, see geom_point.

data
A data frame, see geom_point.

stat
The statistical transformation to use on the data for this layer, as a string, see geom_point.

position
Postion adjustment, see geom_point.

na.rm
See geom_point.

show.legend
See geom_point.

inherit.aes
See geom_point.

...
other arguments passed on to layer. See geom_point.

See also

ggiraph

Examples

# add interactive polygons to a ggplot -------
# create data
ids <- factor(c("1.1", "2.1", "1.2", "2.2", "1.3", "2.3"))

values <- data.frame(
        id = ids,
        value = c(3, 3.1, 3.1, 3.2, 3.15, 3.5) )
positions <- data.frame(
        id = rep(ids, each = 4),
        x = c(2, 1, 1.1, 2.2, 1, 0, 0.3, 1.1, 2.2, 1.1, 1.2, 2.5, 1.1, 0.3,
                0.5, 1.2, 2.5, 1.2, 1.3, 2.7, 1.2, 0.5, 0.6, 1.3),
        y = c(-0.5, 0, 1, 0.5, 0, 0.5, 1.5, 1, 0.5, 1, 2.1, 1.7, 1, 1.5,
                2.2, 2.1, 1.7, 2.1, 3.2, 2.8, 2.1, 2.2, 3.3, 3.2) )

datapoly <- merge(values, positions, by=c("id"))

datapoly$oc = "alert(this.getAttribute(\"data-id\"))"

# create a ggplot -----
gg_poly_1 <- ggplot(datapoly, aes( x = x, y = y ) ) +
        geom_polygon_interactive(aes(fill = value, group = id,
                tooltip = value, data_id = value, onclick = oc))

# display ------
ggiraph(code = {print(gg_poly_1)})

#> Warning: file("") accepte seulement open = "w+" et open = "w+b" : utilisation du premier

#> Error in dirname(to): un argument de type vecteur de caractères est attendu