ggiraph

Create an interactive graphic to be used in a web browser.

Use geom_zzz_interactive to create interactive graphical elements.

Difference from original functions is that the following aesthetics are understood: tooltip, onclick and data_id.

Tooltips can be displayed when mouse is over graphical elements. On click actions can be set with javascript instructions. If id are associated with points, they get animated when mouse is over and can be selected when used in shiny apps.

ggiraph(code, ggobj = NULL, pointsize = 12, width = 0.7, width_svg = 6,
  height_svg = 6, tooltip_extra_css, hover_css, tooltip_opacity = 0.9,
  tooltip_offx = 10, tooltip_offy = 0, zoom_max = 1,
  selection_type = "multiple", selected_css, ...)

Arguments

code

Plotting code to execute

ggobj

ggplot objet to print. argument code will be ignored if this argument is supplied.

pointsize

the default pointsize of plotted text in pixels, default to 12.

width

widget width ratio (0 < width <= 1)

width_svg, height_svg

svg viewbox width and height in inches

tooltip_extra_css

extra css (added to position: absolute;pointer-events: none;) used to customize tooltip area.

hover_css

css to apply when mouse is hover and element with a data-id attribute.

tooltip_opacity

tooltip opacity

tooltip_offx

tooltip x offset

tooltip_offy

tooltip y offset

zoom_max

maximum zoom factor

selection_type

row selection mode ("single", "multiple", "none") when widget is in a Shiny application.

selected_css

css to apply when element is selected (shiny only).

...

arguments passed on to dsvg

Examples

# ggiraph simple example -------
# create dataset
dataset = iris
dataset$tooltip = dataset$Species
dataset$clickjs = paste0("alert(\"",dataset$Species, "\")" )

# plots
gg_point = ggplot(dataset, aes(x = Sepal.Length, y = Petal.Width,
                color = Species, tooltip = tooltip, onclick = clickjs) ) +
        geom_point_interactive()

ggiraph(code = {print(gg_point)})