Get girafe defaults formatting properties

The current formatting properties are automatically applied to every girafe you produce. These default values are returned by this function.

girafe_defaults(name = NULL)

Arguments

name

optional, option's name to return, one of 'fonts', 'opts_sizing', 'opts_tooltip', 'opts_hover', 'opts_hover_key', 'opts_hover_inv', 'opts_hover_theme', 'opts_selection', 'opts_selection_inv', 'opts_selection_key', 'opts_selection_theme', 'opts_zoom', 'opts_toolbar'.

Value

a list containing default values or an element selected with argument name.

See also

Examples

girafe_defaults()
#> $fonts
#> $fonts$sans
#> [1] "DejaVu Sans"
#> 
#> $fonts$serif
#> [1] "DejaVu serif"
#> 
#> $fonts$symbol
#> [1] "DejaVu Sans"
#> 
#> $fonts$mono
#> [1] "DejaVu Sans"
#> 
#> 
#> $opts_sizing
#> $rescale
#> [1] TRUE
#> 
#> $width
#> [1] 1
#> 
#> attr(,"class")
#> [1] "opts_sizing"
#> 
#> $opts_tooltip
#> $css
#> [1] ".tooltip_SVGID_ { padding:5px;background:black;color:white;border-radius:2px;text-align:left; ; position:absolute;pointer-events:none;z-index:999;}"
#> 
#> $placement
#> [1] "doc"
#> 
#> $opacity
#> [1] 0.9
#> 
#> $offx
#> [1] 10
#> 
#> $offy
#> [1] 10
#> 
#> $use_cursor_pos
#> [1] TRUE
#> 
#> $use_fill
#> [1] FALSE
#> 
#> $use_stroke
#> [1] FALSE
#> 
#> $delay_over
#> [1] 200
#> 
#> $delay_out
#> [1] 500
#> 
#> attr(,"class")
#> [1] "opts_tooltip"
#> 
#> $opts_hover
#> $css
#> [1] ".hover_data_SVGID_ { fill:orange;stroke:black;cursor:pointer; }\ntext.hover_data_SVGID_ { stroke:none;fill:orange; }\ncircle.hover_data_SVGID_ { fill:orange;stroke:black; }\nline.hover_data_SVGID_, polyline.hover_data_SVGID_ { fill:none;stroke:orange; }\nrect.hover_data_SVGID_, polygon.hover_data_SVGID_, path.hover_data_SVGID_ { fill:orange;stroke:none; }\nimage.hover_data_SVGID_ { stroke:orange; }"
#> 
#> $reactive
#> [1] TRUE
#> 
#> $nearest_distance
#> NULL
#> 
#> attr(,"class")
#> [1] "opts_hover"
#> 
#> $opts_hover_key
#> $css
#> [1] ".hover_key_SVGID_ { fill:orange;stroke:black;cursor:pointer; }\ntext.hover_key_SVGID_ { stroke:none;fill:orange; }\ncircle.hover_key_SVGID_ { fill:orange;stroke:black; }\nline.hover_key_SVGID_, polyline.hover_key_SVGID_ { fill:none;stroke:orange; }\nrect.hover_key_SVGID_, polygon.hover_key_SVGID_, path.hover_key_SVGID_ { fill:orange;stroke:none; }\nimage.hover_key_SVGID_ { stroke:orange; }"
#> 
#> $reactive
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "opts_hover_key"
#> 
#> $opts_hover_theme
#> $css
#> [1] ".hover_theme_SVGID_ { fill:orange;stroke:black;cursor:pointer; }\ntext.hover_theme_SVGID_ { stroke:none;fill:orange; }\ncircle.hover_theme_SVGID_ { fill:orange;stroke:black; }\nline.hover_theme_SVGID_, polyline.hover_theme_SVGID_ { fill:none;stroke:orange; }\nrect.hover_theme_SVGID_, polygon.hover_theme_SVGID_, path.hover_theme_SVGID_ { fill:orange;stroke:none; }\nimage.hover_theme_SVGID_ { stroke:orange; }"
#> 
#> $reactive
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "opts_hover_theme"
#> 
#> $opts_hover_inv
#> $css
#> [1] ""
#> 
#> attr(,"class")
#> [1] "opts_hover_inv"
#> 
#> $opts_selection
#> $css
#> [1] ".select_data_SVGID_ { fill:red;stroke:black;cursor:pointer; }\ntext.select_data_SVGID_ { stroke:none;fill:red; }\ncircle.select_data_SVGID_ { fill:red;stroke:black; }\nline.select_data_SVGID_, polyline.select_data_SVGID_ { fill:none;stroke:red; }\nrect.select_data_SVGID_, polygon.select_data_SVGID_, path.select_data_SVGID_ { fill:red;stroke:none; }\nimage.select_data_SVGID_ { stroke:red; }"
#> 
#> $type
#> [1] "multiple"
#> 
#> $only_shiny
#> [1] TRUE
#> 
#> $selected
#> character(0)
#> 
#> attr(,"class")
#> [1] "opts_selection"
#> 
#> $opts_selection_inv
#> $css
#> [1] ""
#> 
#> attr(,"class")
#> [1] "opts_selection_inv"
#> 
#> $opts_selection_key
#> $css
#> [1] ".select_key_SVGID_ { fill:red;stroke:black;cursor:pointer; }\ntext.select_key_SVGID_ { stroke:none;fill:red; }\ncircle.select_key_SVGID_ { fill:red;stroke:black; }\nline.select_key_SVGID_, polyline.select_key_SVGID_ { fill:none;stroke:red; }\nrect.select_key_SVGID_, polygon.select_key_SVGID_, path.select_key_SVGID_ { fill:red;stroke:none; }\nimage.select_key_SVGID_ { stroke:red; }"
#> 
#> $type
#> [1] "single"
#> 
#> $only_shiny
#> [1] TRUE
#> 
#> $selected
#> character(0)
#> 
#> attr(,"class")
#> [1] "opts_selection_key"
#> 
#> $opts_selection_theme
#> $css
#> [1] ".select_theme_SVGID_ { fill:red;stroke:black;cursor:pointer; }\ntext.select_theme_SVGID_ { stroke:none;fill:red; }\ncircle.select_theme_SVGID_ { fill:red;stroke:black; }\nline.select_theme_SVGID_, polyline.select_theme_SVGID_ { fill:none;stroke:red; }\nrect.select_theme_SVGID_, polygon.select_theme_SVGID_, path.select_theme_SVGID_ { fill:red;stroke:none; }\nimage.select_theme_SVGID_ { stroke:red; }"
#> 
#> $type
#> [1] "single"
#> 
#> $only_shiny
#> [1] TRUE
#> 
#> $selected
#> character(0)
#> 
#> attr(,"class")
#> [1] "opts_selection_theme"
#> 
#> $opts_zoom
#> $min
#> [1] 1
#> 
#> $max
#> [1] 1
#> 
#> $duration
#> [1] 300
#> 
#> attr(,"class")
#> [1] "opts_zoom"
#> 
#> $opts_toolbar
#> $position
#> [1] "topright"
#> 
#> $pngname
#> [1] "diagram"
#> 
#> $tooltips
#> NULL
#> 
#> $hidden
#> character(0)
#> 
#> $delay_over
#> [1] 200
#> 
#> $delay_out
#> [1] 500
#> 
#> attr(,"class")
#> [1] "opts_toolbar"
#>