Generate interactive grob points

This function can be used to generate interactive grob points.
interactive_points_grob(x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  tooltip = NULL, onclick = NULL, data_id = NULL, pch = 1,
  size = unit(1, "char"), default.units = "native", name = NULL,
  gp = gpar(), vp = NULL)

Arguments

x 	numeric vector or unit object specifying x-values.

y 	numeric vector or unit object specifying y-values.

tooltip 	tooltip associated with points

onclick 	javascript action to execute when point is clicked

data_id 	identifiers to associate with points

pch 	numeric or character vector indicating what sort of plotting symbol to use. See points for the interpretation of these values, and note fill below.

size 	unit object specifying the size of the plotting symbols.

default.units 	string indicating the default units to use if x or y are only given as numeric vectors.

name 	character identifier.

gp 	an R object of class gpar, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings; note that fill (and not bg as in package graphics points) is used to “fill”, i.e., color the background of symbols with pch = 21:25.

vp 	A Grid viewport object (or NULL).