interactive_polygon_grob {ggiraph}	R Documentation

Create interactive polygon grob

Description

The grob is based on polygonGrob(). See the documentation for that function for more details.

Usage

interactive_polygon_grob(...)

Arguments

...
arguments passed to base function, plus any of the interactive_parameters().

Value

An interactive grob object.

Details for interactive_*_grob functions

The interactive parameters can be supplied as arguments in the relevant function and they can be scalar values or vectors depending on params on base function.

See Also

girafe()

[Package ggiraph version 0.7.10 Index]