Generate an Interactive Grob Path

This function can be used to generate an interactive grob path.

interactive_polyline_grob(x = unit(c(0, 1), "npc"), y = unit(c(0, 1),
  "npc"), id = NULL, id.lengths = NULL, tooltip = NULL, onclick = NULL,
  data_id = NULL, default.units = "npc", arrow = NULL, name = NULL,
  gp = gpar(), vp = NULL)

Arguments

x
A numeric vector or unit object specifying x-values.

y
A numeric vector or unit object specifying y-values.

id
A numeric vector used to separate locations in x and y into multiple lines. All locations with the same id belong to the same line.

id.lengths
A numeric vector used to separate locations in x and y into multiple lines. Specifies consecutive blocks of locations which make up separate lines.

tooltip
tooltip associated with polylines

onclick
javascript action to execute when polyline is clicked

data_id
identifiers to associate with polylines

default.units
A string indicating the default units to use if x or y are only given as numeric vectors.

arrow
A list describing arrow heads to place at either end of the line, as produced by the arrow function.

name
A character identifier.

gp
An object of class gpar, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

vp
A Grid viewport object (or NULL).