Generate interactive grob segments

This function can be used to generate interactive grob segments.

interactive_segments_grob(x0 = unit(0, "npc"), y0 = unit(0, "npc"),
  x1 = unit(1, "npc"), y1 = unit(1, "npc"), tooltip = NULL,
  onclick = NULL, data_id = NULL, default.units = "npc", arrow = NULL,
  name = NULL, gp = gpar(), vp = NULL)

Arguments

x0
Numeric indicating the starting x-values of the line segments.

y0
Numeric indicating the starting y-values of the line segments.

x1
Numeric indicating the stopping x-values of the line segments.

y1
Numeric indicating the stopping y-values of the line segments.

tooltip
tooltip associated with segments

onclick
javascript action to execute when segment is clicked

data_id
identifiers to associate with segments

default.units
A string.

arrow
A list describing arrow heads to place at either end of the line segments, as produced by the arrow function.

name
A character identifier.

gp
An object of class gpar.

vp
A Grid viewport object (or NULL).