label_interactive {ggiraph}	R Documentation

Create an interactive label

Description

This function returns an object that can be used as a label via the labs() family of functions or when setting a scale/guide name/title or key label. It passes the interactive parameters to a theme element created via element_text_interactive or via an interactive guide.

Usage

label_interactive(label, ...)

Arguments

label
The text for the label (scalar character)

...
any of the interactive_parameters().

Value

an interactive label object

See Also

interactive_parameters, labeller_interactive()

Examples

library(ggplot2)
library(ggiraph)

gg_jitter <- ggplot(
  mpg, aes(cyl, hwy, group = cyl)) +
  geom_boxplot() +
  labs(title =
         label_interactive(
           "title",
           data_id = "id_title",
           onclick = "alert(\"title\")",
           tooltip = "title" )
  ) +
  theme(plot.title = element_text_interactive())

x <- girafe(ggobj = gg_jitter)
if( interactive() ) print(x)

[Package ggiraph version 0.7.10 Index]