opts_sizing {ggiraph}	R Documentation

Girafe sizing settings

Description

Allows customization of the svg style sizing

Usage

opts_sizing(rescale = TRUE, width = 1)

Arguments

rescale
If FALSE, graphic will not be resized and the dimensions are exactly those of the svg. If TRUE the graphic will be resize to fit its container

width
widget width ratio (0 < width <= 1).

See Also

Other girafe animation options: girafe_options(), opts_hover(), opts_selection(), opts_toolbar(), opts_tooltip(), opts_zoom()

Examples

library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_options(x,
  opts_sizing(rescale = FALSE) )
if( interactive() ) print(x)

[Package ggiraph version 0.7.10 Index]