renderggiraph {ggiraph}	R Documentation

Reactive version of ggiraph object

Description

Makes a reactive version of a ggiraph object for use in Shiny.

This function is now deprecated, users should now use function renderGirafe().

Usage

renderggiraph(expr, env = parent.frame(), quoted = FALSE)

Arguments

expr
An expression that returns a ggiraph() object.

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression

Examples

## Not run: 
if( require(shiny) && interactive() ){
  app_dir <- file.path( system.file(package = "ggiraph"), "examples/shiny" )
  shinyAppDir(appDir = app_dir )
}

## End(Not run)

[Package ggiraph version 0.8.9 Index]