geom_contour_interactive {ggiraph}	R Documentation

Create interactive 2d contours of a 3d surface

Description

These geometries are based on geom_contour() and geom_contour_filled(). See the documentation for those functions for more details.

Usage

geom_contour_interactive(...)

geom_contour_filled_interactive(...)

Arguments

...
arguments passed to base function, plus any of the interactive_parameters().

Details for geom_*_interactive functions

The interactive parameters can be supplied with two ways:

As aesthetics with the mapping argument (via aes()). In this way they can be mapped to data columns and apply to a set of geometries.

As plain arguments into the geom_*_interactive function. In this way they can be set to a scalar value.

See Also

girafe()

Examples

# add interactive contours to a ggplot -------
library(ggplot2)
library(ggiraph)

v <- ggplot(faithfuld, aes(waiting, eruptions, z = density))
p <- v + geom_contour_interactive(aes(
  colour = stat(level),
  tooltip = paste("Level:", stat(level))
))
x <- girafe(ggobj = p)
if (interactive()) print(x)

if (packageVersion("grid") >= numeric_version("3.6")) {
  p <- v + geom_contour_filled_interactive(aes(
    colour = stat(level),
    fill = stat(level),
    tooltip = paste("Level:", stat(level))
  ))
  x <- girafe(ggobj = p)
  if (interactive()) print(x)
}

[Package ggiraph version 0.7.10 Index]