interactive points

The geometry is based on geom_point. See the documentation for those functions for more details.
geom_point_interactive(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)

Arguments

mapping 	The aesthetic mapping, see geom_point.

data 	A data frame, see geom_point.

stat 	The statistical transformation to use on the data for this layer, as a string, see geom_point.

position 	Postion adjustment, see geom_point.

na.rm 	See geom_point.

show.legend 	See geom_point.

inherit.aes 	See geom_point.

... 	other arguments passed on to layer. See geom_point.

Note
The following shapes id 3, 4 and 7 to 14 are composite symbols and should not be used.
See also
ggiraph
Examples

# add interactive points to a ggplot -------
# create dataset
dataset = iris
dataset$tooltip = dataset$Species
dataset$clickjs = paste0("alert(\"",dataset$Species, "\")" )

# plots
gg_point = ggplot(dataset, aes(x = Sepal.Length, y = Petal.Width,
                color = Species, tooltip = tooltip, onclick = clickjs) ) +
        geom_point_interactive()

ggiraph(code = {print(gg_point)})

#> Warning: file("") accepte seulement open = "w+" et open = "w+b" : utilisation du premier

#> Error in dirname(to): un argument de type vecteur de caractères est attendu