Create a ggiraph object

Create an interactive graphic to be used in a web browser.

This function is maintained for backward compatibility reasons, user should now use function girafe() and girafe_options().

ggiraph(
  code,
  ggobj = NULL,
  pointsize = 12,
  width = 0.75,
  width_svg = 6,
  height_svg = 5,
  tooltip_extra_css = NULL,
  hover_css = NULL,
  tooltip_opacity = 0.9,
  tooltip_offx = 10,
  tooltip_offy = 0,
  tooltip_zindex = 999,
  zoom_max = 1,
  selection_type = "multiple",
  selected_css = NULL,
  dep_dir = NULL,
  ...
)

Arguments

code

Plotting code to execute

ggobj

ggplot object to print. Argument code will be ignored if this argument is supplied.

pointsize

the default pointsize of plotted text in pixels, default to 12.

width

widget width ratio (0 < width <= 1).

width_svg, height_svg

The width and height of the graphics region in inches. The default values are 6 and 5 inches. This will define the aspect ratio of the graphic as it will be used to define viewbox attribute of the SVG result.

If you use girafe() in an 'R Markdown' document, we recommend not using these arguments so that the knitr options fig.width and fig.height are used instead.

tooltip_extra_css

extra css (added to position: absolute;pointer-events: none;) used to customize tooltip area.

hover_css

css to apply when mouse is hover and element with a data-id attribute.

tooltip_opacity

tooltip opacity

tooltip_offx

tooltip x offset

tooltip_offy

tooltip y offset

tooltip_zindex

tooltip css z-index, default to 999.

zoom_max

maximum zoom factor

selection_type

row selection mode ("single", "multiple", "none") when widget is in a Shiny application.

selected_css

css to apply when element is selected (shiny only).

dep_dir

Deprecated; the path where the output files are stored. If NULL, the current path for temporary files is used.

...

arguments passed on to dsvg()

Examples

# ggiraph simple example -------
library(ggplot2)
library(ggiraph)

dataset <- structure(list(qsec = c(16.46, 17.02, 18.61, 19.44, 17.02, 20.22
), disp = c(160, 160, 108, 258, 360, 225), carname = c("Mazda RX4",
"Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive", "Hornet Sportabout",
"Valiant"), wt = c(2.62, 2.875, 2.32, 3.215, 3.44, 3.46)), row.names = c("Mazda RX4",
"Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive", "Hornet Sportabout",
"Valiant"), class = "data.frame")
dataset
#>                    qsec disp           carname    wt
#> Mazda RX4         16.46  160         Mazda RX4 2.620
#> Mazda RX4 Wag     17.02  160     Mazda RX4 Wag 2.875
#> Datsun 710        18.61  108        Datsun 710 2.320
#> Hornet 4 Drive    19.44  258    Hornet 4 Drive 3.215
#> Hornet Sportabout 17.02  360 Hornet Sportabout 3.440
#> Valiant           20.22  225           Valiant 3.460

# plots
gg_point = ggplot(data = dataset) +
  geom_point_interactive(aes(x = wt, y = qsec, color = disp,
    tooltip = carname, data_id = carname)) + theme_minimal()

x <- girafe(ggobj = gg_point)
if( interactive() ) print(x)