Create a ggiraph object (deprecated)

Create an interactive graphic to be used in a web browser.

This function is now deprecated, users should now use function girafe().

Usage

ggiraph(
  code,
  ggobj = NULL,
  pointsize = 12,
  width = 0.75,
  width_svg = 6,
  height_svg = 5,
  tooltip_extra_css = NULL,
  hover_css = NULL,
  tooltip_opacity = 0.9,
  tooltip_offx = 10,
  tooltip_offy = 0,
  tooltip_zindex = 999,
  zoom_max = 1,
  selection_type = "multiple",
  selected_css = NULL,
  dep_dir = NULL,
  ...
)

Arguments

code

Plotting code to execute

ggobj

ggplot object to print. Argument code will be ignored if this argument is supplied.

pointsize

the default pointsize of plotted text in pixels, default to 12.

width

widget width ratio (0 < width <= 1).

width_svg, height_svg

The width and height of the graphics region in inches. The default values are 6 and 5 inches. This will define the aspect ratio of the graphic as it will be used to define viewbox attribute of the SVG result.

If you use girafe() in an 'R Markdown' document, we recommend not using these arguments so that the knitr options fig.width and fig.height are used instead.

tooltip_extra_css

extra css (added to position: absolute;pointer-events: none;) used to customize tooltip area.

hover_css

css to apply when mouse is hover and element with a data-id attribute.

tooltip_opacity

tooltip opacity

tooltip_offx

tooltip x offset

tooltip_offy

tooltip y offset

tooltip_zindex

tooltip css z-index, default to 999.

zoom_max

maximum zoom factor

selection_type

row selection mode ("single", "multiple", "none") when widget is in a Shiny application.

selected_css

css to apply when element is selected (shiny only).

dep_dir

Deprecated; the path where the output files are stored. If NULL, the current path for temporary files is used.

...

arguments passed on to dsvg()