girafe_defaults {ggiraph}	R Documentation

Get girafe defaults formatting properties

Description

The current formatting properties are automatically applied to every girafe you produce. These default values are returned by this function.

Usage

girafe_defaults(name = NULL)

Arguments

name
optional, option's name to return, one of 'fonts', 'opts_sizing', 'opts_tooltip', 'opts_hover', 'opts_hover_key', 'opts_hover_inv', 'opts_hover_theme', 'opts_selection', 'opts_selection_inv', 'opts_selection_key', 'opts_selection_theme', 'opts_zoom', 'opts_toolbar'.

Value

a list containing default values or an element selected with argument name.

See Also

Other girafe animation options: girafe_options(), init_girafe_defaults(), opts_hover(), opts_selection(), opts_sizing(), opts_toolbar(), opts_tooltip(), opts_zoom(), set_girafe_defaults()

Examples

girafe_defaults()

[Package ggiraph version 0.8.9 Index]