renderGirafe {ggiraph}	R Documentation

Reactive version of girafe

Description

Makes a reactive version of girafe object for use in Shiny.

Usage

renderGirafe(expr, env = parent.frame(), quoted = FALSE, outputArgs = list())

Arguments

expr
An expression that returns a girafe() object.

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression

outputArgs
A list of arguments to be passed through to the implicit call to girafeOutput() when renderGirafe is used in an interactive R Markdown document.

[Package ggiraph version 0.8.9 Index]