validated_fonts {ggiraph}	R Documentation

List of validated default fonts

Description

Validates and possibly modifies the fonts to be used as default value in a graphic according to the fonts available on the machine. It process elements named "sans", "serif", "mono" and "symbol".

Usage

validated_fonts(fonts = list())

Arguments

fonts
Named list of font names to be aliased with fonts installed on your system. If unspecified, the R default families "sans", "serif", "mono" and "symbol" are aliased to the family returned by match_family().

If fonts are available, the default mapping will use these values:

R family 	 Font on Windows 	 Font on Unix 	 Font on Mac OS

sans 	 Arial 	 DejaVu Sans 	 Helvetica

serif 	 Times New Roman 	 DejaVu serif 	 Times

mono 	 Courier 	 DejaVu mono 	 Courier

symbol 	 Symbol 	 DejaVu Sans 	 Symbol

Value

a named list of validated font family names

See Also

girafe(), dsvg()

Other functions for font management: font_family_exists(), match_family()

Examples

validated_fonts()

[Package ggiraph version 0.8.9 Index]